Add various changes for Heroku deployability
A Procfile was created. A package.json was added to the repo root so
that Heroku sees the repo as a Node.js app. Without it, the Heroku node
buildpack fails. The Heroku node buildpack is forced to use yarn by
puttin yarn in the package.json 'engines' key. package.json also fixes
the node version as Sails is not compatible with some of the latest node
versions.

Other notes:

* npm-run-all isn't used since that's a dev dependency
* The Angular build must done in the root build script!
* Angular and Typescript dependencies need to be prod dependencies
* We need @types/leaflet to be a prod dependency to build the app since
  we do not actually depend on the npm leaflet package.